skb: Rewrote builder from scratch and ton of other improvements.

+++ src/kernel/arch/x86_64/idt.c
@@ -38,28 +38,28 @@ void init_idt()
     idtr.base = (uintptr_t)&idt[0];
     idtr.limit = (uint16_t)sizeof(idt_entry) * 256 - 1;
 
-    idt_set_entry(0, (uint64_t)zero_division_handler, 0x8E);
-    idt_set_entry(1, (uint64_t)debug_handler, 0x8E);
-    idt_set_entry(2, (uint64_t)non_maskable_interrupt_handler, 0x8E);
-    idt_set_entry(3, (uint64_t)breakpoint_handler, 0x8E);
-    idt_set_entry(4, (uint64_t)overflow_handler, 0x8E);
-    idt_set_entry(5, (uint64_t)bound_range_exceeded_handler, 0x8E);
-    idt_set_entry(6, (uint64_t)invalid_opcode_handler, 0x8E);
-    idt_set_entry(7, (uint64_t)device_not_available_handler, 0x8E);
-    idt_set_entry(8, (uint64_t)double_fault_handler, 0x8E);
-    idt_set_entry(10, (uint64_t)invalid_tss_handler, 0x8E);
-    idt_set_entry(11, (uint64_t)segment_not_present_handler, 0x8E);
-    idt_set_entry(12, (uint64_t)stack_fault_handler, 0x8E);
-    idt_set_entry(13, (uint64_t)gpf_handler, 0x8E);
-    idt_set_entry(14, (uint64_t)pf_handler, 0x8E);
-    idt_set_entry(16, (uint64_t)floating_point_error_handler, 0x8E);
-    idt_set_entry(17, (uint64_t)alignment_check_handler, 0x8E);
-    idt_set_entry(18, (uint64_t)machine_check_handler, 0x8E);
-    idt_set_entry(19, (uint64_t)simd_floating_point_handler, 0x8E);
-    idt_set_entry(20, (uint64_t)virtualization_exception_handler, 0x8E);
+    idt_set_entry(0, zero_division_handler, 0x8E);
+    idt_set_entry(1, debug_handler, 0x8E);
+    idt_set_entry(2, non_maskable_interrupt_handler, 0x8E);
+    idt_set_entry(3, breakpoint_handler, 0x8E);
+    idt_set_entry(4, overflow_handler, 0x8E);
+    idt_set_entry(5, bound_range_exceeded_handler, 0x8E);
+    idt_set_entry(6, invalid_opcode_handler, 0x8E);
+    idt_set_entry(7, device_not_available_handler, 0x8E);
+    idt_set_entry(8, double_fault_handler, 0x8E);
+    idt_set_entry(10, invalid_tss_handler, 0x8E);
+    idt_set_entry(11, segment_not_present_handler, 0x8E);
+    idt_set_entry(12, stack_fault_handler, 0x8E);
+    idt_set_entry(13, gpf_handler, 0x8E);
+    idt_set_entry(14, pf_handler, 0x8E);
+    idt_set_entry(16, floating_point_error_handler, 0x8E);
+    idt_set_entry(17, alignment_check_handler, 0x8E);
+    idt_set_entry(18, machine_check_handler, 0x8E);
+    idt_set_entry(19, simd_floating_point_handler, 0x8E);
+    idt_set_entry(20, virtualization_exception_handler, 0x8E);
     // idt_set_entry(0x80, (uint64_t)syscall_handler, 0x8E);
 
-    idt_set_entry(33, (uint64_t)kb_handler, 0x8E);
+    idt_set_entry(33, kb_handler, 0x8E);
 
     remap_pic();
     outb(PIC1_DATA, 0b11111101);
